+++ src/filterConfig.js
@@ -195,13 +195,38 @@ export const instanceIndexes = [
 
 export const browseFiltersConfig = [
   {
+    name: FACETS.SHARED,
+    cql: FACETS_CQL.SHARED,
+    values: [],
+  },
+  {
     name: FACETS.EFFECTIVE_LOCATION,
     cql: FACETS_CQL.EFFECTIVE_LOCATION,
     values: [],
   },
   {
+    name: FACETS.CONTRIBUTORS_SHARED,
+    cql: FACETS_CQL.INSTANCES_SHARED,
+    values: [],
+  },
+  {
+    name: FACETS.CONTRIBUTORS_HELD_BY,
+    cql: FACETS_CQL.INSTANCES_HELD_BY,
+    values: [],
+  },
+  {
     name: FACETS.NAME_TYPE,
     cql: FACETS_CQL.NAME_TYPE,
+    values: [],
+  },
+  {
+    name: FACETS.SUBJECTS_SHARED,
+    cql: FACETS_CQL.INSTANCES_SHARED,
+    values: [],
+  },
+  {
+    name: FACETS.SUBJECTS_HELD_BY,
+    cql: FACETS_CQL.INSTANCES_HELD_BY,
     values: [],
   },
 ];